WARNING : Update to MAME 138u2, added things needed for MESS added later. Big WARNING for this update, everything is changed and there are surely some regressions.
Known issues:
- adding sub devices of cart devices not working
- x1, einstein and osborne needs daisy chain device rewrite
- mm1m6, mm1m7 not in buid since they there is problem UPD7220 device parameter (map)

Regular users please give few days to developers to check and clean drivers before start giving error reports.

git-svn-id: svn://messdev.no-ip.org/mess@8317 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/drivers/ipds.c
@@ -56,7 +56,7 @@ const i8275_interface ipds_i8275_interface = {
 
 static MACHINE_RESET(ipds)
 {
-	cpu_set_reg(devtag_get_device(machine, "maincpu"), I8085_PC, 0x0000);
+	cpu_set_reg(devtag_get_device(machine, "maincpu"), I8085_PC, (UINT64)0x0000);
 }
 
 static VIDEO_UPDATE( ipds )
